requirements.txt
numpy~=2.0.0
matplotlib~=3.9.0
scikit-learn~=1.5.0
black
mealpy~=3.0.1
wandb~=0.17.2
pyswarms~=1.3.0
simanneal~=0.5.0
plotly
pandas
umap-learn
stopit
torch
opencv-python
pytest-cov